File splitup, start of a bounded constraint
