+++ src/caliper/RichAnnotations/CMakeLists.txt
@@ -1,2 +1,3 @@
+add_subdirectory(constraints)
 set(CALIPER_HEADERS ${CALIPER_HEADERS}
         ValidatedAnnotation.h)
