+++ src/caliper/RichAnnotations/ValidatedAnnotation.h
@@ -2,6 +2,8 @@
 #include <iostream>
 #include <functional>
 #include <type_traits>
+#include "constraints/monotonic.h"
+#include "constraints/bounded.h"
 namespace cali{
 
 template<class... ValidatorList>
@@ -155,36 +157,4 @@ class ValidatedAnnotation<Validator, ValidatorList...> : public ValidatedAnnotat
 
 };
 
-template<typename T, typename Comparator>
-struct Monotonic{
-    public:
-    void validateBegin(){}
-    template<typename Q>
-    typename std::enable_if<std::is_same<T,Q>::value,void>::type validateBegin(Q& next){
-        if(Comparator()(next,last)){
-            std::cout<<"ASPOLDE"<<std::endl;
-        }
-    }
-    template<typename Q>
-    typename std::enable_if<!std::is_same<T,Q>::value,void>::type validateBegin(Q& next){
-    }
-    template<typename Q>
-    typename std::enable_if<std::is_same<T,Q>::value,void>::type validateSet(Q& next){
-        if(Comparator()(next,last)){
-            std::cout<<"ASPOLDE"<<std::endl;
-        }
-    }
-    template<typename Q>
-    typename std::enable_if<!std::is_same<T,Q>::value,void>::type validateSet(Q& next){
-    }
-
-    void validateEnd(){
-    }
-    private:
-    T last;
-};
-template<typename T>
-using MonotonicDecreasing = Monotonic<T,std::less_equal<T>>;
-template<typename T>
-using MonotonicIncreasing = Monotonic<T,std::greater_equal<T>>;
 } //end namespace cali
