+++ src/caliper/RichAnnotations/constraints/CMakeLists.txt
@@ -1,0 +1,1 @@
+set(CALIPER_HEADERS ${CALIPER_HEADERS} monotonic.h bounded.h)
